[Tests] Remove duplicates from XASdkTests
I've spent some time reviewing tests in `XASdkTests`, `BuildTest`, and
`BuildTest2`.  Tests from `XASdkTests` have been migrated to other
fixtures or removed where applicable.

 * DotNetBuildXamarinForms has been removed and merged into CheckMonoComponentsMask.

 * DotNetIncremental has been removed and merged into BuildIncrementalAot.

 * DotNetDesignTimeBuild has been renamed to DesignTimeBuildSignAndroidPackage.

 * BenchmarkDotNet has been removed and merged into BuildHasNoWarnings.

 * BuildApplicationWithJavaSource has been removed due to existing coverage of .java file inclusion elsewhere.

 * BuildReleaseApplicationWithNugetPackages has been removed due to existing coverage in GenerateLibraryResourcesTests.

 * BuildAfterAddingNuget has been removed as the same scenario is covered in BuildAfterUpgradingNuget.

 * ValidateJavaVersion has been removed as .NET Android requires JDK 11+, which is not covered by any of the legacy parameters.
